Implemented AWS::CloudFormation::ResourceDefaultVersion
- AWS::CloudFormation::ResourceVersionAlias converted to AWS::CloudFormation::ResourceDefaultVersion
- Implemented as callChain and tests updated
