+++ aws-cloudformation-resourcedefaultversion/.gitignore
