+++ aws-cloudformation-resourcedefaultversion/.rpdk-config
@@ -1,0 +1,15 @@
+{
+    "typeName": "AWS::CloudFormation::ResourceDefaultVersion",
+    "language": "java",
+    "runtime": "java8",
+    "entrypoint": "software.amazon.cloudformation.resourcedefaultversion.HandlerWrapper::handleRequest",
+    "testEntrypoint": "software.amazon.cloudformation.resourcedefaultversion.HandlerWrapper::testEntrypoint",
+    "settings": {
+        "namespace": [
+            "software",
+            "amazon",
+            "cloudformation",
+            "resourcedefaultversion"
+        ]
+    }
+}
