+++ aws-cloudformation-resourcedefaultversion/aws-cloudformation-resourcedefaultversion.json
@@ -1,0 +1,61 @@
+{
+  "typeName": "AWS::CloudFormation::ResourceDefaultVersion",
+  "description": "The default version of a resource that has been registered in the CloudFormation Registry.",
+  "sourceUrl": "https://github.com/aws-cloudformation/aws-cloudformation-resource-providers-cloudformation",
+  "properties": {
+    "Arn": {
+      "description": "The Amazon Resource Name (ARN) of the type version.",
+      "pattern": "^arn:aws[A-Za-z0-9-]{0,64}:cloudformation:[A-Za-z0-9-]{1,64}:([0-9]{12})?:type/resource/.+$",
+      "type": "string"
+    },
+    "TypeName": {
+      "description": "The name of the type being registered.\n\nWe recommend that type names adhere to the following pattern: company_or_organization::service::type.",
+      "pattern": "^[A-Za-z0-9]{2,64}::[A-Za-z0-9]{2,64}::[A-Za-z0-9]{2,64}$",
+      "type": "string"
+    },
+    "VersionId": {
+      "description": "The ID of an existing version of the resource to set as the default.",
+      "pattern": "^[A-Za-z0-9-]{1,128]$",
+      "type": "string"
+    }
+  },
+  "oneOf": [
+    {
+      "required": [
+        "Arn"
+      ]
+    },
+    {
+      "required": [
+        "TypeName",
+        "VersionId"
+      ]
+    }
+  ],
+  "primaryIdentifier": [
+    "/properties/Arn"
+  ],
+  "handlers": {
+    "create": {
+      "permissions": [
+        "cloudformation:SetTypeDefaultVersion"
+      ]
+    },
+    "read": {
+      "permissions": [
+        "cloudformation:DescribeType"
+      ]
+    },
+    "update": {
+      "permissions": [
+        "cloudformation:SetTypeDefaultVersion"
+      ]
+    },
+    "delete": {
+      "permissions": [
+        ""
+      ]
+    }
+  },
+  "additionalProperties": false
+}
