+++ aws-cloudformation-resourcedefaultversion/lombok.config
