+++ aws-cloudformation-resourcedefaultversion/resource-role.yaml
