+++ aws-cloudformation-resourcedefaultversion/src/main/java/software/amazon/cloudformation/resourcedefaultversion/BaseHandlerStd.java
@@ -1,0 +1,33 @@
+package software.amazon.cloudformation.resourcedefaultversion;
+
+import software.amazon.awssdk.services.cloudformation.CloudFormationClient;
+import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
+import software.amazon.cloudformation.proxy.Logger;
+import software.amazon.cloudformation.proxy.ProgressEvent;
+import software.amazon.cloudformation.proxy.ProxyClient;
+import software.amazon.cloudformation.proxy.ResourceHandlerRequest;
+
+public abstract class BaseHandlerStd extends BaseHandler<CallbackContext> {
+  @Override
+  public final ProgressEvent<ResourceModel, CallbackContext> handleRequest(
+    final AmazonWebServicesClientProxy proxy,
+    final ResourceHandlerRequest<ResourceModel> request,
+    final CallbackContext callbackContext,
+    final Logger logger) {
+
+    return handleRequest(
+      proxy,
+      request,
+      callbackContext != null ? callbackContext : CallbackContext.builder().build(),
+      proxy.newProxy(ClientBuilder::getClient),
+      logger
+    );
+  }
+
+  protected abstract ProgressEvent<ResourceModel, CallbackContext> handleRequest(
+    final AmazonWebServicesClientProxy proxy,
+    final ResourceHandlerRequest<ResourceModel> request,
+    final CallbackContext callbackContext,
+    final ProxyClient<CloudFormationClient> proxyClient,
+    final Logger logger);
+}
