+++ aws-cloudformation-resourcedefaultversion/src/main/java/software/amazon/cloudformation/resourcedefaultversion/ClientBuilder.java
@@ -1,0 +1,13 @@
+package software.amazon.cloudformation.resourcedefaultversion;
+
+import software.amazon.awssdk.services.cloudformation.CloudFormationClient;
+import software.amazon.cloudformation.LambdaWrapper;
+
+public class ClientBuilder {
+
+    public static CloudFormationClient getClient() {
+        return CloudFormationClient.builder()
+            .httpClient(LambdaWrapper.HTTP_CLIENT)
+            .build();
+    }
+}
