+++ aws-cloudformation-resourcedefaultversion/src/main/java/software/amazon/cloudformation/resourcedefaultversion/Configuration.java
@@ -1,0 +1,8 @@
+package software.amazon.cloudformation.resourcedefaultversion;
+
+class Configuration extends BaseConfiguration {
+
+    public Configuration() {
+        super("aws-cloudformation-resourcedefaultversion.json");
+    }
+}
