+++ aws-cloudformation-resourcedefaultversion/src/test/java/software/amazon/cloudformation/resourcedefaultversion/DeleteHandlerTest.java
@@ -1,0 +1,42 @@
+package software.amazon.cloudformation.resourcedefaultversion;
+
+import software.amazon.awssdk.services.cloudformation.CloudFormationClient;
+import software.amazon.cloudformation.proxy.OperationStatus;
+import software.amazon.cloudformation.proxy.ProgressEvent;
+import software.amazon.cloudformation.proxy.ResourceHandlerRequest;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
+import software.amazon.cloudformation.test.AbstractMockTestBase;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+@ExtendWith(MockitoExtension.class)
+public class DeleteHandlerTest extends AbstractMockTestBase<CloudFormationClient> {
+    final DeleteHandler handler = new DeleteHandler();
+
+    protected DeleteHandlerTest() {
+        super(CloudFormationClient.class);
+    }
+
+    @Test
+    public void handleRequest_SimpleSuccess() {
+        final ResourceModel model = ResourceModel.builder().build();
+
+        final ResourceHandlerRequest<ResourceModel> request = ResourceHandlerRequest.<ResourceModel>builder()
+            .desiredResourceState(model)
+            .build();
+
+        final ProgressEvent<ResourceModel, CallbackContext> response
+            = handler.handleRequest(proxy, request, null, loggerProxy);
+
+        assertThat(response).isNotNull();
+        assertThat(response.getStatus()).isEqualTo(OperationStatus.SUCCESS);
+        assertThat(response.getCallbackContext()).isNull();
+        assertThat(response.getCallbackDelaySeconds()).isEqualTo(0);
+        assertThat(response.getResourceModel()).isEqualTo(request.getDesiredResourceState());
+        assertThat(response.getResourceModels()).isNull();
+        assertThat(response.getMessage()).isNull();
+        assertThat(response.getErrorCode()).isNull();
+    }
+}
