--- aws-cloudformation-resourceversionalias/.rpdk-config
@@ -1,15 +1,0 @@
-{
-    "typeName": "AWS::CloudFormation::ResourceVersionAlias",
-    "language": "java",
-    "runtime": "java8",
-    "entrypoint": "software.amazon.cloudformation.resourceversionalias.HandlerWrapper::handleRequest",
-    "testEntrypoint": "software.amazon.cloudformation.resourceversionalias.HandlerWrapper::testEntrypoint",
-    "settings": {
-        "namespace": [
-            "software",
-            "amazon",
-            "cloudformation",
-            "resourceversionalias"
-        ]
-    }
-}
