--- aws-cloudformation-resourceversionalias/aws-cloudformation-resourceversionalias.json
@@ -1,74 +1,0 @@
-{
-  "typeName": "AWS::CloudFormation::ResourceVersionAlias",
-  "description": "The default version of a resource that has been registered in the CloudFormation Registry.",
-  "sourceUrl": "https://github.com/aws-cloudformation/aws-cloudformation-resource-providers-cloudformation",
-  "properties": {
-    "Arn": {
-      "description": "The Amazon Resource Name (ARN) of the type version.",
-      "pattern": "^arn:aws[A-Za-z0-9-]{0,64}:cloudformation:[A-Za-z0-9-]{1,64}:([0-9]{12})?:type/.+$",
-      "type": "string"
-    },
-    "DefaultVersionId": {
-      "description": "The ID of an existing version of the resource to set as the default.",
-      "maxLength": 128,
-      "minLength": 1,
-      "pattern": "^[A-Za-z0-9-]+$",
-      "type": "string"
-    },
-    "TypeName": {
-      "description": "The name of the type being registered.\n\nWe recommend that type names adhere to the following pattern: company_or_organization::service::type.",
-      "maxLength": 196,
-      "minLength": 10,
-      "pattern": "^[A-Za-z0-9]{2,64}::[A-Za-z0-9]{2,64}::[A-Za-z0-9]{2,64}$",
-      "type": "string"
-    }
-  },
-  "oneOf": [
-    {
-      "required": [
-        "DefaultVersionId",
-        "Arn"
-      ]
-    },
-    {
-      "required": [
-        "DefaultVersionId",
-        "Type",
-        "TypeName"
-      ]
-    }
-  ],
-  "required": [
-    "DefaultVersionId",
-    "TypeName"
-  ],
-  "readOnlyProperties": [
-    "/properties/Arn"
-  ],
-  "primaryIdentifier": [
-    "/properties/Arn"
-  ],
-  "handlers": {
-    "create": {
-      "permissions": [
-        "cloudformation:SetTypeDefaultVersion"
-      ]
-    },
-    "read": {
-      "permissions": [
-        "cloudformation:DescribeType"
-      ]
-    },
-    "update": {
-      "permissions": [
-        "cloudformation:SetTypeDefaultVersion"
-      ]
-    },
-    "delete": {
-      "permissions": [
-        ""
-      ]
-    }
-  },
-  "additionalProperties": false
-}
