--- aws-cloudformation-resourceversionalias/src/main/java/software/amazon/cloudformation/resourceversionalias/CallbackContext.java
@@ -1,12 +1,0 @@
-package software.amazon.cloudformation.resourceversionalias;
-
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@Builder
-public class CallbackContext {
-
-}
