--- aws-cloudformation-resourceversionalias/src/main/java/software/amazon/cloudformation/resourceversionalias/ClientBuilder.java
@@ -1,10 +1,0 @@
-package software.amazon.cloudformation.resourceversionalias;
-
-import software.amazon.awssdk.services.cloudformation.CloudFormationClient;
-
-public class ClientBuilder {
-    public static CloudFormationClient getClient() {
-        return CloudFormationClient.builder()
-            .build();
-    }
-}
