--- aws-cloudformation-resourceversionalias/src/main/java/software/amazon/cloudformation/resourceversionalias/Configuration.java
@@ -1,8 +1,0 @@
-package software.amazon.cloudformation.resourceversionalias;
-
-class Configuration extends BaseConfiguration {
-
-    public Configuration() {
-        super("aws-cloudformation-resourceversionalias.json");
-    }
-}
