--- aws-cloudformation-resourceversionalias/src/main/java/software/amazon/cloudformation/resourceversionalias/DeleteHandler.java
@@ -1,27 +1,0 @@
-package software.amazon.cloudformation.resourceversionalias;
-
-import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
-import software.amazon.cloudformation.proxy.Logger;
-import software.amazon.cloudformation.proxy.ProgressEvent;
-import software.amazon.cloudformation.proxy.OperationStatus;
-import software.amazon.cloudformation.proxy.ResourceHandlerRequest;
-
-public class DeleteHandler extends BaseHandler<CallbackContext> {
-
-    @Override
-    public ProgressEvent<ResourceModel, CallbackContext> handleRequest(
-        final AmazonWebServicesClientProxy proxy,
-        final ResourceHandlerRequest<ResourceModel> request,
-        final CallbackContext callbackContext,
-        final Logger logger) {
-
-        final ResourceModel model = request.getDesiredResourceState();
-
-        // Delete is a no-op for this resource type as a default version must always exist for a type
-
-        return ProgressEvent.<ResourceModel, CallbackContext>builder()
-            .resourceModel(model)
-            .status(OperationStatus.SUCCESS)
-            .build();
-    }
-}
