--- aws-cloudformation-resourceversionalias/src/test/java/software/amazon/cloudformation/resourceversionalias/UpdateHandlerTest.java
@@ -1,81 +1,0 @@
-package software.amazon.cloudformation.resourceversionalias;
-
-import org.mockito.ArgumentMatchers;
-import software.amazon.cloudformation.exceptions.ResourceNotFoundException;
-import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
-import software.amazon.cloudformation.proxy.Logger;
-import software.amazon.cloudformation.proxy.OperationStatus;
-import software.amazon.cloudformation.proxy.ProgressEvent;
-import software.amazon.cloudformation.proxy.ResourceHandlerRequest;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-
-@ExtendWith(MockitoExtension.class)
-public class UpdateHandlerTest {
-    private UpdateHandler handler;
-
-    @Mock
-    private AmazonWebServicesClientProxy proxy;
-
-    @Mock
-    private Logger logger;
-
-    @BeforeEach
-    public void setup() {
-        handler = new UpdateHandler();
-        proxy = mock(AmazonWebServicesClientProxy.class);
-        logger = mock(Logger.class);
-    }
-
-    @Test
-    public void handleRequest_SimpleSuccess() {
-        final ResourceModel model = ResourceModel.builder().build();
-
-        final ResourceHandlerRequest<ResourceModel> request = ResourceHandlerRequest.<ResourceModel>builder()
-            .desiredResourceState(model)
-            .build();
-
-        final ProgressEvent<ResourceModel, CallbackContext> response
-            = handler.handleRequest(proxy, request, null, logger);
-
-        assertThat(response).isNotNull();
-        assertThat(response.getStatus()).isEqualTo(OperationStatus.SUCCESS);
-        assertThat(response.getCallbackContext()).isNull();
-        assertThat(response.getCallbackDelaySeconds()).isEqualTo(0);
-        assertThat(response.getResourceModel()).isEqualTo(request.getDesiredResourceState());
-        assertThat(response.getResourceModels()).isNull();
-        assertThat(response.getMessage()).isNull();
-        assertThat(response.getErrorCode()).isNull();
-    }
-
-    @Test
-    public void handleRequest_NotFound() {
-        final ResourceModel resourceModel = ResourceModel.builder()
-            .arn("arn:aws:cloudformation:us-west-2:123456789012:type/resource/AWS-Demo-Resource/00000001")
-            .typeName("AWS::Demo::Resource")
-            .build();
-
-        when(proxy.injectCredentialsAndInvokeV2(
-            ArgumentMatchers.any(),
-            ArgumentMatchers.any())
-        )
-            .thenThrow(new ResourceNotFoundException("AWS::CloudFormation::ResourceVersionAlias", resourceModel.getArn()));
-
-        final ResourceHandlerRequest<ResourceModel> request = ResourceHandlerRequest.<ResourceModel>builder()
-            .desiredResourceState(resourceModel)
-            .build();
-
-        assertThatThrownBy(() -> handler.handleRequest(proxy, request, null, logger))
-            .hasNoCause()
-            .hasMessage("Resource of type 'AWS::CloudFormation::ResourceVersionAlias' with identifier '{\"/properties/Arn\":\"arn:aws:cloudformation:us-west-2:123456789012:type/resource/AWS-Demo-Resource/00000001\"}' was not found.")
-            .isExactlyInstanceOf(ResourceNotFoundException.class);
-    }
-}
